Fix: unknown pragma 'clang optimize' error when g++ is used for building VIXL tests
In Linaro ART CI, VIXL tests are run as a part of CI. The tests are built
both by clang++ and g++.

The patch "Disable optimizations for instructions-aarch32.h in a test"
https://android-review.googlesource.com/c/platform/external/vixl/+/611524
added the pragma 'clang optimize'. As g++ does not support the pragma
and '-Werror=unknown-pragmas' is used CI fails trying to build tests.

This patch adds ifdef to fix the error.

Test: CXX=g++ scons -C ./external/vixl all mode=debug std=c++98
Test: mmma external/vixl finishes in 4m vs without the pragma
Change-Id: Iaaed1f2f13985a202b9152964e1c251b9e332aaf