Export include dirs for use by art
Also changed vixl-test-runner to depend on libvixld instead of building
the code to be tested from source again. That avoided having to add
"src" to the local_include_dirs specified in "vixl-test-runner" because
the export_include_dirs specified in the defaults is not supported on
cc_test_host and so was ignored.

Test: m checkbuild
Bug: 134379140
Change-Id: Ia77a393e4fe8f11f64534b47015f1b0fc9de940c