Add GetScratchVRegisterList to help ART to move to new API
In the latest upstream VIXL GetScratchFPRegisterList gets replaced with
GetScratchVRegisterList and is not available any more. This causes issues
to move ART to the new API.

The current version of GetScratchVRegisterList is temporarily added till
the ART is moved to the new VIXL API and the upstream version of
GetScratchVRegisterList is being integrated.

Test: mma test-art-host-vixl
Change-Id: If527928cc0cc1e9de38502717211007374de3d6e